rename: accordion 컴포넌트 폴더 이름 수정
