+++ src/app/components/accordion/AccordionBody.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react'
 
-import { useAccordionContext } from '@components/accordion-temp/AccordionContext'
+import { useAccordionContext } from '@components/accordion/AccordionContext'
 
 export default function AccordionBody({ children }: PropsWithChildren) {
   const { isOpen } = useAccordionContext()
